Restructure snapshot generator pkg
