+++ pkg/snapshot/generator/postgres/data/pg_schema_table_parser.go
